feat(discoveryDetailsPlaceholders): Refactored remaining renderers into separate files
