+++ packages/frontend/src/features/Study/StudyDetails/StudyItems/Renderers/index.ts
@@ -6,3 +6,8 @@ export { default as LabeledMultipleTextField } from './LabeledMultipleTextField'
 export { default as LabeledSingleTextField } from './LabeledSingleTextField';
 export { default as LabeledSingleLinkField } from './LabeledSingleLinkField';
 export { default as LabeledParagraph } from './LabeledParagraph';
+export { default as UnlabeledMultipleLinkField } from './UnlabeledMultipleLinkField';
+export { default as LabeledYearOfBirthRestricted } from './LabeledYearOfBirthRestricted';
+export { default as AccessDescriptor } from './AccessDescriptor';
+export { default as DataDownloadListField } from './DataDownloadListField';
+export { default as RenderDetailTags } from './RenderDetailTags';
