+++ packages/frontend/src/features/Study/StudyDetails/StudyItems/StudyItems.tsx
@@ -1,14 +1,11 @@
-import React, { ReactElement, useId } from 'react';
-import Link from 'next/link';
-import { isArray, toString } from 'lodash';
+import React, { ReactElement } from 'react';
+import { isArray } from 'lodash';
 import { JSONPath } from 'jsonpath-plus';
-import { Alert, Text } from '@mantine/core';
 import {
   StudyDetailsField,
   StudyResource,
   StudyTabTagField,
 } from '../../types';
-import { accessibleFieldName, AccessLevel } from '../../../../utils';
 import { RenderTagsCell } from '../../TableRenderers/CellRenderers';
 import {
   FieldRendererFunction,
@@ -16,8 +13,7 @@ import {
   StudyDetailsRenderer,
   StudyFieldRendererFactory,
 } from './RendererFactory';
-import { fieldNameToLabel, JSONObject, JSONValue } from '@gen3/core';
-import DataDownloadList from '../DataDownload/DataDownloadList';
+import { fieldNameToLabel, JSONValue } from '@gen3/core';
 import {
   BlockTextField,
   LabeledSingleLinkField,
@@ -26,136 +22,12 @@ import {
   LabeledMultipleTextField,
   LabeledSingleTextField,
   LabeledParagraph,
+  UnlabeledMultipleLinkField,
+  LabeledYearOfBirthRestricted,
+  AccessDescriptor,
+  DataDownloadListField,
+  RenderDetailTags,
 } from './Renderers';
-
-const linkFieldOnly = (linkValue: string, _?: string) => (
-  <Link href={linkValue} target="_blank" rel="noreferrer">
-    {linkValue}
-  </Link>
-);
-
-/**
- * Represents a link field component that generates a hyperlink with an optional text label.
- *
- * @param linkValue - The target URL for the hyperlink.
- * @returns A JSX element containing the generated hyperlink.
- */
-const linkField = (linkValue: string) => (
-  <Link href={linkValue} target="_blank" rel="noreferrer">
-    <Text c="utility.0" className="underline">
-      {linkValue}
-    </Text>
-  </Link>
-);
-
-const subHeading = (text: string) => (
-  <h3 className="discovery-subheading">{text}</h3>
-);
-
-interface LinkWithTitle {
-  title: string;
-  link: string;
-}
-
-const unlabeledMultipleLinkField = (
-  fieldData: JSONValue,
-  fieldName?: string,
-) => {
-  if (!isArray(fieldData) || fieldData.length === 0) return <React.Fragment />;
-  const links = fieldData[0] as unknown as LinkWithTitle[];
-  return (
-    <div className="flex flex-col" key={`${fieldName}-links`}>
-      {links.map((link) => LabeledSingleLinkField(link.link, link.title))}
-    </div>
-  );
-};
-
-const labeledYearOfBirthRestricted: FieldRendererFunction = (
-  fieldValue: JSONValue,
-  fieldLabel?: string,
-) => {
-  let stringFieldValue = '';
-  if (typeof fieldValue === 'number') {
-    stringFieldValue = fieldValue.toLocaleString();
-  } else if (typeof fieldValue !== 'string') return <React.Fragment />;
-
-  stringFieldValue = fieldValue as string;
-
-  let displayContent;
-  if (
-    typeof stringFieldValue === 'string' &&
-    !isNaN(Number(stringFieldValue)) &&
-    Number(stringFieldValue) < 1935
-  ) {
-    displayContent = '1935';
-  } else if (isArray(stringFieldValue)) {
-    displayContent = stringFieldValue
-      .map((item) => {
-        if (
-          typeof item === 'string' &&
-          !isNaN(Number(item)) &&
-          Number(item) < 1935
-        ) {
-          return '1935';
-        }
-        return item;
-      })
-      .join(', ');
-  } else {
-    displayContent = stringFieldValue;
-  }
-
-  return (
-    <div
-      className={discoveryFieldStyle}
-      key={`study-details-${fieldLabel}-${displayContent}`}
-    >
-      {label(fieldLabel)} {textField(displayContent)}
-    </div>
-  );
-};
-
-const dataDownloadList: FieldRendererFunction = (
-  resource: JSONValue,
-  _: string | undefined,
-) => {
-  return (
-    <>
-      <DataDownloadList data={resource as JSONObject} />
-    </>
-  );
-};
-
-const accessDescriptor: FieldRendererFunction = (
-  resource: JSONValue,
-  _: string | undefined,
-) => {
-  if (
-    resource === null ||
-    typeof resource !== 'object' /*
-    COMMENTING THIS OUT WILL BE IMPLEMENTED UNTIL HP-2378
-    ||
-    !(accessibleFieldName in resource) */
-  ) {
-    return <></>;
-  }
-
-  if (
-    (resource as JSONObject)[accessibleFieldName] === AccessLevel.ACCESSIBLE
-  ) {
-    return <Alert color="green">You have access to this study.</Alert>;
-  }
-  if (
-    (resource as JSONObject)[accessibleFieldName] === AccessLevel.UNACCESSIBLE
-  ) {
-    return <Alert color="red">You do not have access to this study.</Alert>;
-  }
-  return (
-    <Alert color="yellow">
-      This study does not include data access authorization details.
-    </Alert>
-  );
-};
 
 const formatResourceValuesWhenNestedArray = (
   resourceFieldValue: string[],
@@ -167,37 +39,6 @@ const formatResourceValuesWhenNestedArray = (
     return resourceFieldValue[0].join(', ');
   }
   return resourceFieldValue;
-};
-
-const renderDetailTags: FieldRendererFunction = (
-  fieldValue: JSONValue,
-  _label: string | undefined,
-  fieldConfig?: Record<string, any>,
-): ReactElement => {
-  //TODO - fix this type
-  const resource = fieldValue as StudyResource;
-
-  if (fieldConfig === undefined) {
-    return <React.Fragment />;
-  }
-  if (fieldConfig?.categories === undefined) {
-    return <React.Fragment />;
-  }
-
-  if (fieldConfig.contentType === 'tags') {
-    const tags = fieldConfig.categories
-      ? (resource.tags || []).filter((tag) =>
-          fieldConfig.categories?.includes(tag.category),
-        )
-      : resource.tags;
-
-    return (
-      <div key={`detail-tag-${fieldConfig.field}`}>
-        {RenderTagsCell({ value: tags })}
-      </div>
-    );
-  }
-  return <React.Fragment />;
 };
 
 /**
@@ -281,7 +122,6 @@ export const createFieldRendererElement = (
       )
         return studyFieldRenderer(resourceFieldValue, label, field.params);
   }
-
   return null;
 };
 
@@ -292,18 +132,18 @@ const DefaultGen3StudyDetailsFieldsRenderers: Record<
   text: { default: LabeledSingleTextField },
   string: {
     default: LabeledSingleTextField,
-    yearOfBirthRestricted: labeledYearOfBirthRestricted,
+    yearOfBirthRestricted: LabeledYearOfBirthRestricted,
   },
-  dataDownloadList: { default: dataDownloadList },
+  dataDownloadList: { default: DataDownloadListField },
   link: { default: LabeledSingleLinkField },
   textList: { default: LabeledMultipleTextField },
   linkList: {
     default: LabeledMultipleLinkField,
-    linkWithTitle: unlabeledMultipleLinkField,
+    linkWithTitle: UnlabeledMultipleLinkField,
   },
   block: { default: BlockTextField },
-  accessDescriptor: { default: accessDescriptor },
-  tags: { default: renderDetailTags },
+  accessDescriptor: { default: AccessDescriptor },
+  tags: { default: RenderDetailTags },
   number: { default: LabeledNumberField },
   paragraphs: { default: LabeledParagraph },
 };
